Remove broken upgrade wizard (#5283)
