--- core/src/main/resources/jenkins/install/UpgradeWizard/client-scripts_bg.properties
@@ -1,31 +1,0 @@
-# The MIT License
-#
-# Bulgarian translation: Copyright (c) 2016, Alexander Shopov <ash@kambanaria.org>
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-
-# \u0020to get the new features
-msg.after=\
- , \u0437\u0430 \u0434\u0430 \u0441\u0435 \u0441\u0434\u043e\u0431\u0438\u0435\u0442\u0435 \u0441 \u043d\u043e\u0432\u0438\u0442\u0435 \u0432\u044a\u0437\u043c\u043e\u0436\u043d\u043e\u0441\u0442\u0438
-# Upgrade now
-msg.link=\
- \u041e\u0431\u043d\u043e\u0432\u044f\u0432\u0430\u043d\u0435
-# Welcome to Jenkins 2!\u0020
-msg.before=\
- \u0414\u043e\u0431\u0440\u0435 \u0434\u043e\u0448\u043b\u0438 \u0432 Jenkins 2!\u0020
