--- core/src/main/resources/jenkins/install/UpgradeWizard/client-scripts_fr.properties
@@ -1,3 +1,0 @@
-msg.before=Bienvenue sur Jenkins {0} !\u0020
-msg.link=Installer les nouvelles fonctionnalit\u00E9s
-msg.after=\u0020pour terminer la mise \u00E0 jour
